renderScript/requirements.txt
requests
python-dotenv
pymongo[srv]
